Added Support Devexpress 17.2.3
