--- FarsiLibrary.Win.DevExpress.17.1/Properties/licenses.licx
@@ -1,2 +1,0 @@
-DevExpress.XtraEditors.Repository.RepositoryItemDateEdit, DevExpress.XtraEditors.v17.1, Version=17.1.4.0, Culture=neutral, PublicKeyToken=b88d1754d700e49a
-DevExpress.XtraEditors.DateEdit, DevExpress.XtraEditors.v17.1, Version=17.1.4.0, Culture=neutral, PublicKeyToken=b88d1754d700e49a
