+++ Solution Items/CommonAssemblyInfo.cs
@@ -21,4 +21,4 @@ using FarsiLibrary.Localization;
 [assembly: InternalsVisibleTo("FarsiLibrary.WebDemo, PublicKey=00240000048000009400000006020000002400005253413100040000010001000bd06c5f2139bb49786f0f662880bd6e984a64b40bfbedbd39f5893396c22b88366c9e1586bd2c9758e28063266d1b4cd521b89b9eb5f6332b4bf780fbb94926bd31416c0f4f2c4c4d95bc300241602447fdbf2b17cda756c51c4831f6f314fab2296bf319d9dbd2ef3a05dda68fd8495642aa5d376bfdd6fbfce193448893c4")]
 [assembly: InternalsVisibleTo("FarsiLibrary.Win.DevExpress.15.2, PublicKey=00240000048000009400000006020000002400005253413100040000010001000bd06c5f2139bb49786f0f662880bd6e984a64b40bfbedbd39f5893396c22b88366c9e1586bd2c9758e28063266d1b4cd521b89b9eb5f6332b4bf780fbb94926bd31416c0f4f2c4c4d95bc300241602447fdbf2b17cda756c51c4831f6f314fab2296bf319d9dbd2ef3a05dda68fd8495642aa5d376bfdd6fbfce193448893c4")]
 [assembly: InternalsVisibleTo("FarsiLibrary.Win.DevExpress.16.1, PublicKey=00240000048000009400000006020000002400005253413100040000010001000bd06c5f2139bb49786f0f662880bd6e984a64b40bfbedbd39f5893396c22b88366c9e1586bd2c9758e28063266d1b4cd521b89b9eb5f6332b4bf780fbb94926bd31416c0f4f2c4c4d95bc300241602447fdbf2b17cda756c51c4831f6f314fab2296bf319d9dbd2ef3a05dda68fd8495642aa5d376bfdd6fbfce193448893c4")]
-[assembly: InternalsVisibleTo("FarsiLibrary.Win.DevExpress.17.1, PublicKey=00240000048000009400000006020000002400005253413100040000010001000bd06c5f2139bb49786f0f662880bd6e984a64b40bfbedbd39f5893396c22b88366c9e1586bd2c9758e28063266d1b4cd521b89b9eb5f6332b4bf780fbb94926bd31416c0f4f2c4c4d95bc300241602447fdbf2b17cda756c51c4831f6f314fab2296bf319d9dbd2ef3a05dda68fd8495642aa5d376bfdd6fbfce193448893c4")]
+[assembly: InternalsVisibleTo("FarsiLibrary.Win.DevExpress.17.2, PublicKey=00240000048000009400000006020000002400005253413100040000010001000bd06c5f2139bb49786f0f662880bd6e984a64b40bfbedbd39f5893396c22b88366c9e1586bd2c9758e28063266d1b4cd521b89b9eb5f6332b4bf780fbb94926bd31416c0f4f2c4c4d95bc300241602447fdbf2b17cda756c51c4831f6f314fab2296bf319d9dbd2ef3a05dda68fd8495642aa5d376bfdd6fbfce193448893c4")]
